requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==21.2.0
mysqlclient==2.2.0
num2words==0.5.13
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
